Import the updated BLAKE2 implementation from CPython
Replace the outdated BLAKE2 implementation with the one used in CPython.
While it does not seem to be the newest upstream implementation, it is
certainly newer than the one currently used and according to user
reports it does not exhibit segfaults with optimizations on newer
hardware (issue #13). Most importantly, its structure resembles the one
used in pyblake2 much more, so the upgrade is easier.

While at it, try to preserve the upstream code structure as much as
possible to ease future maintenance. In particular, we should be able to
easily backport various fixes from CPython. The only difference
in the impl/ directory is removal of <stdint.h> includes that was
historically done to improve compatibility with old MSVC versions.

Two new top-level 'implementation wrapper' files were added to match
CPython whose purpose is switching between the reference and optimized
SSE2+ implementation. The files were also used to provide int*_n
definitions for stdint.h/MSVC compatibility.

The support for explicitly switching BLAKE2_COMPRESS* implementation
has been removed temporarily in favor of upstream CFLAGS-based
switching. They will be reintroduced in an optional manner.