Reintroduce the support for selecting optimized version
Restore the support for selecting optimized version of BLAKE2
implementation. However, make the choice optional and default to
autodetecting based on CFLAGS. This ensures that normally (e.g. when
installing via pip/virtualenv) the best version supported by the target
is used, while making it possible for an expert user to override it
willingly.